fix(dg): relax the requirements for existing PRs
After introducing the commit action, we allow customizing of the
beginning of the PR description (which includes the commit description).
This conflicts with the pre-existing check for existing (pun not
intended) PRs on dist-git.

Relaxing this condition to match the title, source & target branch with
the user should be enough to not duplicate PRs.

Signed-off-by: Matej Focko <mfocko@redhat.com>