Be able to specify bugs resolved by an update
Add an argument to sync_release that takes list of bugs. These will
be then added to the changelog in the end as "- Resolves {bug}".
In case of using autochangelog, these bugs will be referenced in the end
of the commit message. Besides these default behaviours, Packit will
provide the PACKIT_RESOLVED_BUGS env variable to commit-message and changelog-entry
actions.

Related to #1920