Introduce new job trigger: "ignore"
In some cases it could be valuable for users to create a job that
is ignored by Packit (read: it's not executed). Typical use case
is when users want to define a template job that could be used in
real test jobs using the yaml format.
E.g.:
```
      jobs:
    - &my-abstract-job
      job: tests
      trigger: ignore
      ... other generic repeating values ...

    - my-real-job
      <<: *my-abstract-job
      trigger: pull_request
      ...test specific values...
```

This is especially useful for projects with many similar jobs defined
in the .packit.yaml configuration file. Using templates users can
make their .packit.yaml config file much shorter, which reduces also
maintainance burden.

The original problem was, that each job requires a defined trigger,
so it was not possible to create a template that is not executed
at all. For this purposes, it's possible to specify "ignore" trigger,
which will not be executed by packit.

This can be used also to disable some jobs temporarily, e.g. when
they are wanted in future, however due to an issue/changes
(e.g. environment change, infra) they will be failing for a period
of time.

Issue: #2231