fix(env): expand and patch variables for Sandcastle
Apparently expanding the common environment to other actions results in
unexpected cloning of the non-existing repos. These variables haven't
been provided before **and** they are handled in a special way because
of the Sandcastle.

Fill the variables for both Sandcastle and co-Sandcastle command
handlers and then patch them for Sandcastle, if needed. This should
avoid unnecessary cloning of the dist-git in the upstream Koji build
where it has been discovered.

Discovered by @lbarcziova on an upstream Koji build
Fixes packit/packit-service#2444

Signed-off-by: Matej Focko <mfocko@redhat.com>