Get rid of (unused) fedora.client.AuthError import
The code for BodhiClient cannot throw fedora.client.AuthError. The code that
did that was removed in commit 32fa9f4fcd850558b8f9f024989dd151bb068452,
released in bodhi 5.7.0. Also, old bodhi clients cannot work anymore,
because OpenID auth no longer works with our current Bodhi deployments and
will not be "fixed" to do so, and bodhi-client < 6 cannot authenticate anyway.

Getting rid of the import will allow the whole dependency on python-fedora
to be dropped, which we need to do, because that package is going away.