fix(cli): correctly type the resolved bugs
During the changes in #2428 we have incorrectly assumed the type of the
bugs that are passed to the ‹-b› (or ‹--resolve-bug›) switch.

The previous assumption expected that the bugs are represented by simple
number (integer) whereas they are represented in a bit more complex form,
such as ‹rhbz#1234› (Bugzilla) or even ‹PROJECT-1234› (Jira).

Revert such assumption to accept plain strings as it makes the ‹-b›
swich unusable with such precondition.

Signed-off-by: Matej Focko <mfocko@redhat.com>