Always include the list of bugs in `BodhiClient.save()` kwargs
It seems there is a bug/feature in Bodhi and when editing an existing
update and not supplying the `bugs` parameter, Bodhi returns a failure
however the only non-updated property of the update is the list of
bugs.

Always including the list of bugs, even if empty, should prevent this.

Signed-off-by: Nikola Forró <nforro@redhat.com>