Initiate dist-git repo after spec file is present in `cockpit_ostree` fixture
The spec file is copied to the dist-git repo after it is initialized,
which means it isn't commited to git and eventually gets removed,
but that doesn't make the tests fail because the related `Specfile`
instance keeps its content in memory even though the file no longer
exists. That however changes once the `Specfile` instance gets properly
reloaded when needed and the tests start to fail. Fix that.

Signed-off-by: Nikola Forró <nforro@redhat.com>