ci(tmt): lower tier of ‹session-recording›
Based on the observation of the runs on the Testing Farm, it appears that
the ‹session-recording› plan takes under a minute to run, therefore it
could be considered a quick test to be run (as opposed to the full test
suite).

Additionally reverse-dependency tests being run from the ‹ogr› filter
only for tiers 0 and 1 which can result in the session-recording not
being run and actually breaking the test suite.

Given the speed and reverse-dependency points above, it makes sense
to run the ‹session-recording› tests all the time when submitting PRs
to the ‹ogr›.

Also filter tests for ‹session-recording› plan based on the ‹requre› tag
instead of the tier as it pulls in the ‹/tests/full›.

Signed-off-by: Matej Focko <mfocko@redhat.com>