fix(env): do not treat all env variables as paths
Subtle detail sneaked in with #2589, not all variables that we set are
paths, e.g., there are also package names. Trying to make those paths in
Sandcastle relative to the current working directory is impossible, duh,
they are just names, not actual paths.

Therefore process the paths separately, also factor out in the process,
so it cannot be so easily overlooked (for the full list of people that
didn't notice, have a look below lol).

As for the specfile path itself, that was a cause for this change, I am
not sure what is the expected behavior… Leaving it «as is» for now, docs
suggest it should be relative path within the upstream repo, but we're
talking about ‹sync-release› actions here, so there's also
‹pull-from-upstream› that doesn't require an actual upstream repo…

We can change it later, if needed…

Honorary members of the blind society:
• @kayodegigz
• @LecrisUT
• @mfocko
• @lbarcziova

o7

Fixes #2626

Signed-off-by: Matej Focko <mfocko@packit.dev>