Add check for pre-commit hook during packit init
"Packit init" in CLI now checks for the existence of a .git repo and
the presence of pre-commit hook in .git/hooks/pre-commit. If either is
not found, PackitException is raised and user is prompted to initialize
current repo as a git repo and is asked to install pre-commit hook.

This is to enforce the installation of pre-commit hook early on
to avoid future unnecessary frustration of users.