Add proper parsing of file and remove check rebase from configuration
1. Remove check-rebase check from pre-commit config generation as it
is only meant to be used internally.

2. Fix improper generation of pre-commit configuration check in
configuration file. Previously, packit-specific checks were only
appended at the end of the file, regardless of the file structure
and its validity, which could result in issues. In this new version,
the file is parsed and checked for validity before inserting thw
packit-specific check (validate-config).
The file is also checked for whether this check is already defined
in the user's file, in which case, the check is not added for
the second time. This check is also added if the configuration file
is empty. If there are issues found with the user's configuration
file, an error is raised.

3. Rename append_precommit_config to generate_precommit_config.