ci(packit): adjust create-archive
As it has been discovered by @betulependule in #2641, it appears that
there is an issue with creating an archive with hatch.

During closer inspection, I have discovered that the path to the archive,
which we *guess*, is not correct, therefore switch to the `hatch build`
and capture the output of the build, then we should be guaranteed that
the last line of the output is a path to the created archive.

Signed-off-by: Matej Focko <mfocko@packit.dev>